+++ public/app/features/trails/auto-query/query-generators/baseQuery.ts
@@ -1,4 +1,4 @@
-import { VAR_METRIC_EXPR, VAR_FILTERS_WITH_CURLY_EXPR, VAR_OTEL_JOIN_QUERY_EXPR } from '../../../shared';
+import { VAR_METRIC_EXPR, VAR_FILTERS_WITH_CURLY_EXPR, VAR_OTEL_JOIN_QUERY_EXPR } from '../../shared';
 
 const GENERAL_BASE_QUERY = `${VAR_METRIC_EXPR}${VAR_FILTERS_WITH_CURLY_EXPR}`;
 const GENERAL_RATE_BASE_QUERY = `rate(${GENERAL_BASE_QUERY}[$__rate_interval])`;
